Clean up core.py and sync the client and the server versions. (#308)
The purpose of core.py is to share data structures between client and server; nothing else. Therefore:
- Delete dataframe diffing code; it was never finished, and is not used. It stays in git history if we want it back.
- Move CommandException to command.py. Its purpose is to report high-level user-friendly errors from commands; nothing outside of command.py should raise it.
- Move BuildException to build.py. Same as above; nothing outside of build should be raising it.
- Move yaml parsing into build.py
- Move package parsing helpers into command.py. They deal with parsing data from the CLI, so likely not useful elsewhere.

For now, there are still two different core.py files; will unify them in a follow-up PR.
