+++ compiler/quilt/tools/command.py
@@ -32,13 +32,12 @@ from six.moves.urllib.parse import urlparse, urlunparse
 from tqdm import tqdm
 
 from .build import (build_package, build_package_from_contents, generate_build_file,
-                    generate_contents, BuildException)
+                    generate_contents, BuildException, exec_yaml_python, load_yaml)
 from .const import DEFAULT_BUILDFILE, LATEST_TAG
 from .core import (hash_contents, find_object_hashes, PackageFormat, TableNode, FileNode, GroupNode,
-                   decode_node, encode_node, exec_yaml_python, CommandException, diff_dataframes,
-                   load_yaml)
+                   decode_node, encode_node)
 from .hashing import digest_file
-from .store import PackageStore, parse_package, parse_package_extended, VALID_NAME_RE
+from .store import PackageStore, StoreException, VALID_NAME_RE
 from .util import BASE_DIR, FileWithReadProgress, gzip_compress
 from . import check_functions as qc
 
@@ -71,6 +70,69 @@ CONTENT_RANGE_RE = re.compile(r'^bytes (\d+)-(\d+)/(\d+)$')
 LOG_TIMEOUT = 3  # 3 seconds
 
 VERSION = pkg_resources.require('quilt')[0].version
+
+
+class CommandException(Exception):
+    """
+    Exception class for all command-related failures.
+    """
+    pass
+
+def parse_package_extended(name):
+    hash = version = tag = None
+    try:
+        if ':' in name:
+            name, versioninfo = name.split(':', 1)
+            if ':' in versioninfo:
+                info = versioninfo.split(':', 1)
+                if len(info) == 2:
+                    if 'version'.startswith(info[0]):
+                        # usr/pkg:v:<string>  usr/pkg:version:<string>  etc
+                        version = info[1]
+                    elif 'tag'.startswith(info[0]):
+                        # usr/pkg:t:<tag>  usr/pkg:tag:<tag>  etc
+                        tag = info[1]
+                    elif 'hash'.startswith(info[0]):
+                        # usr/pkg:h:<hash>  usr/pkg:hash:<hash>  etc
+                        hash = info[1]
+                    else:
+                        raise CommandException("Invalid versioninfo: %s." % info)
+                else:
+                    # usr/pkg:hashval
+                    hash = versioninfo
+        team, owner, pkg, subpath = parse_package(name, allow_subpath=True)
+    except ValueError:
+        pkg_format = 'owner/package_name/path[:v:<version> or :t:tag or :h:hash]'
+        raise CommandException("Specify package as %s." % pkg_format)
+    return owner, pkg, subpath, hash, version, tag
+
+def parse_package(name, allow_subpath=False):
+    try:
+        values = name.split(':', 1)
+        team = values[0] if len(values) > 1 else None
+
+        values = values[-1].split('/')
+        # Can't do "owner, pkg, *subpath = ..." in Python2 :(
+        (owner, pkg), subpath = values[:2], values[2:]
+        if not owner or not pkg:
+            # Make sure they're not empty.
+            raise ValueError
+        if subpath and not allow_subpath:
+            raise ValueError
+
+    except ValueError:
+        pkg_format = '[team:]owner/package_name/path' if allow_subpath else '[team:]owner/package_name'
+        raise CommandException("Specify package as %s." % pkg_format)
+
+    try:
+        PackageStore.check_name(team, owner, pkg, subpath)
+    except StoreException as ex:
+        raise CommandException(str(ex))
+
+    if allow_subpath:
+        return team, owner, pkg, subpath
+    return team, owner, pkg
+
 
 _registry_url = None
 
@@ -378,26 +440,6 @@ def generate(directory, outfilename=DEFAULT_BUILDFILE):
         raise CommandException(str(builderror))
 
     print("Generated build-file %s." % (buildfilepath))
-
-def diff_node_dataframe(package, nodename, dataframe):
-    """
-    compare two dataframes and print the result
-
-    WIP: find_node_by_name() doesn't work yet.
-    TODO: higher level API: diff_two_files(filepath1, filepath2)
-    TODO: higher level API: diff_node_file(file, package, nodename, filepath)
-    """
-    raise NotImplementedError()
-
-    team, owner, pkg = parse_package(package)
-    pkgobj = PackageStore.find_package(team, owner, pkg)
-    if pkgobj is None:
-        raise CommandException("Package {owner}/{pkg} not found.".format(owner=owner, pkg=pkg))
-    node = pkgobj.find_node_by_name(nodename)
-    if node is None:
-        raise CommandException("Node path not found: {}".format(nodename))
-    quilt_dataframe = pkgobj.get_obj(node)
-    return diff_dataframes(quilt_dataframe, dataframe)
 
 def check(path=None, env='default'):
     """
