+++ compiler/quilt/tools/store.py
@@ -7,7 +7,7 @@ import re
 from shutil import rmtree
 
 from .const import DEFAULT_TEAM, PACKAGE_DIR_NAME
-from .core import FileNode, RootNode, TableNode, CommandException
+from .core import FileNode, RootNode, TableNode
 from .package import Package, PackageException
 from .util import BASE_DIR, sub_dirs, sub_files
 
@@ -307,62 +307,3 @@ class PackageStore(object):
             os.remove(self.object_path(obj))
             removed.append(obj)
         return removed
-
-########################################
-# Helper Functions
-########################################
-
-def parse_package_extended(name):
-    hash = version = tag = None
-    try:
-        if ':' in name:
-            name, versioninfo = name.split(':', 1)
-            if ':' in versioninfo:
-                info = versioninfo.split(':', 1)
-                if len(info) == 2:
-                    if 'version'.startswith(info[0]):
-                        # usr/pkg:v:<string>  usr/pkg:version:<string>  etc
-                        version = info[1]
-                    elif 'tag'.startswith(info[0]):
-                        # usr/pkg:t:<tag>  usr/pkg:tag:<tag>  etc
-                        tag = info[1]
-                    elif 'hash'.startswith(info[0]):
-                        # usr/pkg:h:<hash>  usr/pkg:hash:<hash>  etc
-                        hash = info[1]
-                    else:
-                        raise CommandException("Invalid versioninfo: %s." % info)
-                else:
-                    # usr/pkg:hashval
-                    hash = versioninfo
-        team, owner, pkg, subpath = parse_package(name, allow_subpath=True)
-    except ValueError:
-        pkg_format = 'owner/package_name/path[:v:<version> or :t:tag or :h:hash]'
-        raise CommandException("Specify package as %s." % pkg_format)
-    return owner, pkg, subpath, hash, version, tag
-
-def parse_package(name, allow_subpath=False):
-    try:
-        values = name.split(':', 1)
-        team = values[0] if len(values) > 1 else None
-
-        values = values[-1].split('/')
-        # Can't do "owner, pkg, *subpath = ..." in Python2 :(
-        (owner, pkg), subpath = values[:2], values[2:]
-        if not owner or not pkg:
-            # Make sure they're not empty.
-            raise ValueError
-        if subpath and not allow_subpath:
-            raise ValueError
-
-    except ValueError:
-        pkg_format = '[team:]owner/package_name/path' if allow_subpath else '[team:]owner/package_name'
-        raise CommandException("Specify package as %s." % pkg_format)
-
-    try:
-        PackageStore.check_name(team, owner, pkg, subpath)
-    except StoreException as ex:
-        raise CommandException(str(ex))
-
-    if allow_subpath:
-        return team, owner, pkg, subpath
-    return team, owner, pkg
